requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
anyio==4.2.0
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
base58==2.1.1
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==7.1.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
duckdb==0.9.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
Faker==19.13.0
fastapi==0.108.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.1
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.216
langchainplus-sdk==0.0.20
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numexpr==2.8.8
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==0.28.1
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandasai==1.5.19
parso==0.8.3
Pillow==10.1.0
platformdirs==4.1.0
posthog==3.1.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
pyparsing==3.1.1
pypdf==3.17.4
PyPDF2==3.0.1
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
scipy==1.12.0
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.24
stack-data==0.6.3
starlette==0.32.0.post1
streamlit==1.29.0
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.25.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.12
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0